[MATLAB] emit autodoc for functions
Octave allows storing documentatino via texinfo in the .oct file, but matlab doesn't.
Therefore, the code was changed to write documentation strings in the .m files.
This is currently only done for functions.

As a cleanup, the "docs" hash has been removed. The autodoc strings are now
stored as attributes attached to the original node (e.g. "matlab:synopsis").