[MATLAB] initial attempt to get test-suite to work
Added matlab to configure.ac etc and checked in a few *runme.m.
Currently I get warnings about try-catch syntax (on matlab 2012a) but
it does work.

Examples/test-suite/matlab/Makefile.in attempts to disable tests with
directors as these are known to fail. That will have to be removed
later.